Refactored position of react components to give space for the lobby
