+++ frontend/src/components/Game/MainPlayer/MainPlayer.jsx
