+++ frontend/src/components/Game/MainPlayer/MainPlayerName.jsx
